booking-service/requirements.txt
flask
kafka-python
flask-cors
mysql-connector-python
